Try the original grader before the rewritten one
The failures that were left in the numpy removal were not about the solution;
they were about check.py. A rewritten grader imports solution_ref by name and
raises ModuleNotFoundError, or compares a returned list against an array and
reports an error of infinity, or stops producing JSON at all. Eleven of fifty
failures in one window came back that way.

The original grader usually still works: a reference written out by hand
iterates its argument whether that argument is a list or an array. So the task
is now verified with check.py untouched first, and the rewritten one is only
tried if that fails. The window after the change went from 31 of 50 to 17 of 22.

Includes the tasks and units verified since the last commit.

+++ tasks/rwa-measure-kept-token-count-and-memory-from-compression-ratio/solution_ref.py
@@ -1,6 +1,3 @@
-import numpy as np
-import math
-
 def measure_kept_tokens_and_memory(compression_ratio: float,
                                    seq_len: int,
                                    full_bytes: int) -> tuple[int, float]:
@@ -21,10 +18,10 @@ def measure_kept_tokens_and_memory(compression_ratio: float,
     -------
     kept_tokens : int
         Rounded number of tokens that are retained: round((1 - r) * seq_len).
-    memory_saved : float64
+    memory_saved : float
         Amount of KV memory saved: r * full_bytes.
     """
     val = (1.0 - compression_ratio) * seq_len
     kept = int(round(val))
-    saved = np.float64(compression_ratio * full_bytes)
+    saved = float(compression_ratio * full_bytes)
     return kept, saved
